Refactor sensors.py
Fixes bug where sensors stop working after reloading integration
