+++ custom_components/hass_pontos/utils.py
@@ -15,3 +15,27 @@ async def fetch_data(ip, url_list):
                     LOGGER.error(f"Failed to fetch data: HTTP {response.status}")
     return data
 
+# Parsing sensor data
+def parse_data(data, sensor):
+    """Process, format, and validate sensor data."""
+    if data is None:
+        return None
+    _data = data.get(sensor._endpoint, None)
+
+    # Apply format replacements if format_dict is present
+    if sensor._format_dict is not None and _data is not None:
+        for old, new in sensor._format_dict.items():
+            _data = _data.replace(old, new)
+
+    # Translate alarm codes if code_dict is present
+    if sensor._code_dict is not None and _data is not None:
+        _data = sensor._code_dict.get(_data, _data)
+
+    # Scale sensor data if scale is present
+    if sensor._scale is not None and _data is not None:
+        try:
+            _data = float(_data) * sensor._scale
+        except (ValueError, TypeError):
+            pass
+
+    return _data
